+++ src/main/java/com/programmers/wallet/controller/WalletController.java
@@ -1,12 +1,9 @@
 package com.programmers.wallet.controller;
 
-import com.programmers.customer.Customer;
 import com.programmers.view.View;
-import com.programmers.voucher.voucher.Voucher;
 import com.programmers.wallet.service.WalletService;
 import org.springframework.stereotype.Component;
 
-import java.util.List;
 import java.util.UUID;
 
 import static com.programmers.message.Message.VOUCHER_CUSTOMER_ID;
@@ -32,27 +29,6 @@ public class WalletController {
         UUID voucherUUID = UUID.fromString(voucherId);
 
         walletService.assignVoucher(customerUUID, voucherUUID);
-    }
-
-    public void showCustomerVouchers() {
-        view.printMessage(VOUCHER_CUSTOMER_ID);
-        String customerId = view.getUserCommand();
-        UUID customerUUID = UUID.fromString(customerId);
-
-        List<Voucher> vouchers = walletService.searchVouchersByCustomerId(customerUUID);
-
-        for (Voucher voucher : vouchers) {
-            view.printVoucher(voucher);
-        }
-    }
-
-    public void findVoucherOwner() {
-        view.printMessage(VOUCHER_ID);
-        String voucherId = view.getUserCommand();
-        UUID voucherUUID = UUID.fromString(voucherId);
-
-        Customer customer = walletService.searchCustomerByVoucherId(voucherUUID);
-        view.printCustomer(customer);
     }
 
     public void delete() {
