+++ src/main/resources/views/customers.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
+          name="viewport">
+    <meta content="ie=edge" http-equiv="X-UA-Compatible">
+    <link crossorigin="anonymous" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css"
+          integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" rel="stylesheet">
+
+    <title>Document</title>
+</head>
+<body class="container-fluid">
+<h1>KDT Spring App</h1>
+<p th:text="'The time on the server is' + ${serverTime}"></p>
+<table class="table table-striped table-hover">
+    <thead>
+    <tr>
+        <th scope="col">ID</th>
+        <th scope="col">Name</th>
+        <th scope="col">Email</th>
+        <th scope="col">CreateAt</th>
+        <th scope="col">LastLogin</th>
+    </tr>
+    </thead>
+    <tbody>
+    <tr th:each="customer : ${customers}" th:object="${customers}">
+        <th scope="row" th:text="${customer.customer_id}"></th>
+        <th scope="row" th:text="${customer.name}"></th>
+        <th scope="row" th:text="${customer.email}"></th>
+        <th scope="row" th:text="${customer.createAt}"></th>
+        <th scope="row" th:text="${customer.lastLoginAt}"></th>
+    </tr>
+    </tbody>
+</table>
+
+<div class="d-grid gap-2 d-md-flex justify-content-md-end">
+    <a btn btn-primary me-md-2 th:href="@{/customers/new}" type="button">Button</a>
+</div>
+</body>
+</html>
