+++ src/main/resources/views/new-customers.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
+          name="viewport">
+    <meta content="ie=edge" http-equiv="X-UA-Compatible">
+    <link crossorigin="anonymous" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css"
+          integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" rel="stylesheet">
+
+    <title>
+        KDT - New Customer
+    </title>
+</head>
+<body class="container-fluid">
+<h1>KDT Spring App</h1>
+<h2>New Customer</h2>
+<form method="post" th:action="@{/customers/new}">
+    <div class="mb-3">
+        <label class="form-label" for="exampleInputEmail1">Email address</label>
+        <input aria-describedby="emailHelp" class="form-control" id="exampleInputEmail1" name="email" type="email">
+    </div>
+    <div class="mb-3">
+        <label class="form-label" for="exampleInputPassword1">Name</label>
+        <input class="form-control" id="exampleInputPassword1" name="name" type="text">
+    </div>
+    <button class="btn btn-primary" type="submit">Submit</button>
+</form>
+</body>
+</html>
